+++ cli/src/core/index.ts
@@ -307,7 +307,7 @@ export async function runCli(): Promise<CliResults> {
         { value: "nextjs", name: "Next.js" },
         { value: "payload", name: "Payload CMS" },
       ],
-      default: !defaultOptions.flags.backend,
+      default: defaultOptions.flags.backend,
     })
     if (project.backend === "payload") {
       project.databaseProvider = await select({
@@ -316,7 +316,10 @@ export async function runCli(): Promise<CliResults> {
           { value: "sqlite", name: "SQLite" },
           { value: "postgresql", name: "PostgreSQL" },
         ],
-        default: !defaultOptions.flags.dbProvider,
+        default: defaultOptions.flags.dbProvider as Exclude<
+          DatabaseProvider,
+          "mysql"
+        >,
       })
     }
 
@@ -332,7 +335,7 @@ export async function runCli(): Promise<CliResults> {
           { value: "authjs", name: "Auth.js" },
           { value: "betterAuth", name: "Better Auth" },
         ],
-        default: !defaultOptions.flags.authentication,
+        default: defaultOptions.flags.authentication,
       })
       project.orm = await select({
         message: "What database ORM would you like to use?",
@@ -341,7 +344,7 @@ export async function runCli(): Promise<CliResults> {
           { value: "prisma", name: "Prisma" },
           { value: "drizzle", name: "Drizzle" },
         ],
-        default: !defaultOptions.flags.orm,
+        default: defaultOptions.flags.orm,
       })
       if (project.orm !== "none") {
         project.databaseProvider = await select({
@@ -351,7 +354,7 @@ export async function runCli(): Promise<CliResults> {
             { value: "mysql", name: "MySQL" },
             { value: "postgresql", name: "PostgreSQL" },
           ],
-          default: !defaultOptions.flags.dbProvider,
+          default: defaultOptions.flags.dbProvider,
         })
       }
     }
@@ -362,7 +365,7 @@ export async function runCli(): Promise<CliResults> {
         { value: "eslint/prettier", name: "ESLint/Prettier" },
         { value: "biome", name: "Biome" },
       ],
-      default: !defaultOptions.flags.linter,
+      default: defaultOptions.flags.linter,
     })
     if (!cliResults.flags.git) {
       project.git = await confirm({