+++ deprecated/v1/src/components/mdx/figure.tsx
@@ -1,9 +1,0 @@
-import { type ComponentProps } from "react"
-
-import { cn } from "@/lib/utils"
-
-export function Figure(props: ComponentProps<"figure">) {
-  const { className, ...rest } = props
-
-  return <figure className={cn(className)} {...rest} />
-}