+++ docs/v2/src/lib/config.ts
@@ -16,7 +16,7 @@ export const siteConfig = {
     "https://opengraph.b-cdn.net/production/images/dc4fc423-b479-4749-affc-7d5cde0f5587.png?token=b2Fi8FhjWFLaTtm3lT7_CRZRMf3AjDD78LlkOStxOGk&height=675&width=1200&expires=33294259508",
   links: {
     github: "https://github.com/SlickYeet/create-lx2-app",
-    discord: "https://link.lx2.dev/discord",
+    discord: "https://hub.lx2.dev/discord",
   },
   navItems: [
     {